agents/requirements.txt
# Agno Framework - Sistema Multiagente
agno>=0.1.0

# Twilio WhatsApp Integration
twilio>=9.0.0

# FastAPI para el servidor
fastapi>=0.109.0
uvicorn>=0.27.0

# HTTP client
httpx>=0.26.0

# Environment variables
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Groq for LLM (ultra-fast inference)
groq>=0.4.0

# Pydantic for data validation
pydantic>=2.5.0

# Async support
asyncio-throttle>=1.0.2
